chore: refactored symbols editor and disabled default exports
