+++ .eslintrc.js
@@ -6,6 +6,7 @@ module.exports = {
   rules: {
     'react/react-in-jsx-scope': 0,
     'react/prop-types': 0,
-    'react/jsx-props-no-spreading': 0
+    'react/jsx-props-no-spreading': 0,
+    'import/prefer-default-export': 0
   }
 }
