+++ src/components/NavBar.tsx
@@ -94,5 +94,3 @@ export const NavBar: FC = () => {
     </>
   );
 };
-
-export default NavBar;
