+++ src/hooks/useIsMobile.ts
@@ -2,5 +2,3 @@ import { mobileBreakpoint } from '../config/constants';
 import { useMatchMedia } from './useMatchMedia';
 
 export const useIsMobile = () => useMatchMedia(mobileBreakpoint);
-
-export default useIsMobile;
