+++ src/hooks/useLocalStorageList.ts
@@ -211,5 +211,3 @@ export function useLocalStorageList<T>(
     getKey,
   };
 }
-
-export default useLocalStorageList;
