+++ src/hooks/usePrevious.ts
@@ -9,5 +9,3 @@ export const usePrevious = <T>(value: T): T | undefined => {
 
   return ref.current;
 };
-
-export default usePrevious;
