+++ src/screens/Recipe.tsx
@@ -126,5 +126,3 @@ export const RecipeScreen: FC = () => {
     </RecipeProvider>
   );
 };
-
-export default RecipeScreen;
